added config system, added trait class
