+++ project/plugins.sbt
@@ -1,1 +1,2 @@
 logLevel := Level.Warn
+addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.14.3")
