+++ src/main/resources/application.conf
@@ -1,0 +1,2 @@
+spark.master=""
+spark.appName="preprocessor"
